+++ packages/gov-ui-kit/CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `InputFileAvatar` component
+-   Implement `InputFileAvatar` and `Dropdown` components
+-   Ref property handling on `Button` component
 
 ## [1.0.13] - 2024-02-14